backend/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
redis[hiredis]==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyyaml==6.0.1
prometheus-fastapi-instrumentator==6.1.0
